Move examples out of src directory
The examples is not code that should be executable, but merly to be used
as an example. So moved it out of the autoloaded src directory into it's
own examples directory.
